Fix base on categories
The base is now fixed.  The category app was inhereting base.html from transactions.  Therefore if running different apps and you want different bases for whatever reason, you need to have different filenames for the base.html.  Or change this to have a master base file if it can function on all apps.